added all the declarations from summary to a separate service
