+++ src/app/shared/components/summary-tile/summary-tile.component.html
@@ -43,7 +43,7 @@
         <div
           *ngIf="actions.can_add_approver"
           class="summary-tile--approvers-add-approver"
-          (click)="openApproverListDialog()"
+          (click)="openApproverListDialogFunction()"
         >
           <img [src]="'../../../../assets/svg/fy-edit-gradient.svg'" class="approver--icon" />
         </div>
